refactor(settings): M-1c — extract home-view sub-components from SettingsView.swift (UI-002)
Audit M-1c: final structural pass on SettingsView.swift. Moves the home-screen
sub-components to a single shared file under Settings/v2/Components/.

Extracted (Components/SettingsHomeViews.swift, ~110 lines):
- SettingsHomeHeader (was kept in SettingsView through M-1b for Scaffold reuse;
  now logically grouped with other home views)
- SettingsCategoryCard
- FlexibleBadgeRow
- SettingsBadgeView

SettingsView.swift drops 401 → 294 lines (~27% further reduction).
Total since M-1a start: 1170 → 294 lines (~75% reduction).

The remaining 294 lines are: enums + types (lines 1-68), the SettingsView
coordinator + its review-destination helpers (lines 70-292). This is the
right size for the coordinator — no further decomposition needed.

Visibility bumped on SettingsSummaryBadge from private → internal so
SettingsCategoryCard can reference it from the new file.

No behaviour changes — pure structural decomposition. Build green.

Plan: docs/superpowers/plans/2026-04-19-m1-settings-v3-decomposition.md
